revamped character movements
