+++ js/updateDOM.js
@@ -1,0 +1,16 @@
+//DOM Elements
+// const scene = document.getElementById("scene");
+// const characterScene = document.getElementById("char-grid");
+// const projectilesScene = document.getElementById("proj-grid");
+// const timeDisplay = document.getElementById("time");
+// const defeatDisplay = document.getElementById("defeat");
+
+// export function updateDOM(elements) {
+//   for (const element in elements) {
+//     let elementDOM = document.createElement("div");
+//     elementDOM.classList.add(element.image);
+//     element.elementDOM.style.gridColumn = `${element.currentPosition.x}`;
+//     element.elementDOM.style.gridRow = `${element.currentPosition.y}`;
+//     element.parentDOMElement.appendChild(element.elementDOM.domElement);
+//   }
+// }
