+++ js/utils.js
@@ -1,0 +1,2 @@
+export const randomPosition = max =>
+  Math.floor(Math.random() * Math.floor(max));
