+++ style.css
@@ -6,8 +6,9 @@ body {
   margin: 0;
   background: black;
 }
+
 div#scene {
-  background: url("./landscape.gif") no-repeat;
+  background: url("./ui/landscape.gif") no-repeat;
   background-size: contain;
   width: 1080px;
   height: 520px;
@@ -15,9 +16,11 @@ div#scene {
   z-index: -1;
   overflow: hidden;
 }
+
 audio {
   display: none;
 }
+
 #char-grid {
   display: grid;
   width: 1080px;
@@ -27,6 +30,7 @@ audio {
   position: absolute;
   z-index: 10;
 }
+
 #proj-grid {
   display: grid;
 
@@ -37,6 +41,7 @@ audio {
   position: absolute;
   z-index: 0;
 }
+
 .wrapper {
   display: flex;
   justify-content: center;
@@ -44,6 +49,7 @@ audio {
   width: 100%;
   height: 700px;
 }
+
 .projectile {
   background: url("./projectiles/fireball.png");
   background-repeat: no-repeat;
@@ -52,6 +58,7 @@ audio {
   width: 50px;
   transform: rotate(90deg);
 }
+
 .characterRunning {
   height: 55px;
   width: 45px;
@@ -59,6 +66,7 @@ audio {
   background-position-y: -70px;
   animation: runnin 0.5s steps(4) infinite;
 }
+
 .characterJumping {
   height: 55px;
   width: 45px;
@@ -81,17 +89,26 @@ audio {
   grid-row: 3 / 4;
   animation: idle 1s steps(3) infinite;
 }
+
 .reversed {
   transform: scaleX(-1);
+  height: 55px;
+  width: 45px;
+  background: url("./sprites/sprite.gif");
+  background-position-y: -70px;
+  animation: runnin 0.5s steps(4) infinite;
 }
+
 .lives {
   position: absolute;
   top: 15%;
 }
+
 .life {
   height: 20px;
   width: 20px;
 }
+
 .time-elapsed {
   top: 0px;
   position: absolute;
@@ -99,18 +116,22 @@ audio {
   display: flex;
   background: red;
 }
+
 #defeat {
   position: absolute;
   text-align: center;
   color: yellow;
   display: none;
 }
+
 #defeat h1 {
   font-size: 10rem;
 }
+
 #defeat p {
   font-size: 5rem;
 }
+
 #defeat p.nope {
   font-size: 0.5rem;
 }
@@ -123,6 +144,7 @@ audio {
     background-position-x: -210px;
   }
 }
+
 @keyframes jumping {
   from {
     background-position-x: -10px;
